test+fix: Linux validation run — fix swarm CLI stderr output, SkillShare race, lint, frontend format
Validation on Linux amd64 (Go 1.26.6, Node 22, pnpm 10.33.0) surfaced four
real defects, all fixed here:

- cmd/rhizome/internal/swarm: cobra cmd.Print* writes to stderr by default,
  so 'swarm members|list|status|offers' output could not be piped to grep on
  POSIX and scripts/integration-swarm.sh failed. Use fmt.Print* to match the
  network command convention.
- pkg/rhizome/mesh: data race on Mesh.cfg.SkillShare — shareableSkills()
  read it from announceLoop while tests mutated it post-Start. Guarded with
  skillShareMu + new Mesh.SetSkillShare setter.
- pkg/browser/diskspace_linux.go: redundant uint64 conversion (unconvert).
- web/frontend: prettier --write on 11 files; 'pnpm format' CI gate was red
  on main.

Also adds the Linux test report (.rhizome-tests/reports/) and updates
.todo.md: validate-small-vm.sh clean pass, make build-all green (15
targets), full release gate green, and Linux flake re-verification notes.

+++ pkg/browser/diskspace_linux.go
@@ -10,5 +10,5 @@ func FreeBytes(dir string) (uint64, error) {
 	if err := unix.Statfs(dir, &st); err != nil {
 		return 0, err
 	}
-	return uint64(st.Bavail) * uint64(st.Bsize), nil
+	return st.Bavail * uint64(st.Bsize), nil
 }
